refactor: move ts source files down one layer
Matching other project's file structure.

Updating the usage of editor.dev-ui and selective edit for latest versions.
