+++ src/ts/api/githubApi.ts
@@ -31,7 +31,7 @@ import {
   UrlConfig,
   WorkspaceData,
   WorkspacePublishConfig,
-} from '@blinkk/editor.dev-ui/dist/src/editor/api';
+} from '@blinkk/editor.dev-ui/dist/editor/api';
 import {
   FileNotFoundError,
   ProjectTypeStorageComponent,
