+++ src/ts/projectType/projectType.ts
@@ -2,14 +2,14 @@ import {
   EditorFileData,
   FileData,
   ProjectData,
-} from '@blinkk/editor.dev-ui/dist/src/editor/api';
+} from '@blinkk/editor.dev-ui/dist/editor/api';
 import {
   GetFileRequest,
   GetProjectRequest,
   SaveFileRequest,
   UploadFileRequest,
 } from '../api/api';
-import {FilterComponent} from '@blinkk/editor.dev-ui/dist/src/utility/filter';
+import {FilterComponent} from '@blinkk/editor.dev-ui/dist/utility/filter';
 import {ProjectTypeStorageComponent} from '../storage/storage';
 import express from 'express';
