+++ src/ts/utility/yamlSchemas.ts
@@ -5,8 +5,8 @@ import {
   YamlTypeComponent,
   YamlTypeConstructor,
 } from './yamlConvert';
-import {DeepObject} from '@blinkk/selective-edit/dist/src/utility/deepObject';
-import {DeepWalk} from '@blinkk/editor.dev-ui/dist/src/utility/deepWalk';
+import {DeepObject} from '@blinkk/selective-edit/dist/utility/deepObject';
+import {DeepWalk} from '@blinkk/editor.dev-ui/dist/utility/deepWalk';
 import {ProjectTypeStorageComponent} from '../storage/storage';
 import yaml from 'js-yaml';
